Split Ensemble1 to files
