+++ run_reproduce.sh
@@ -1,4 +1,4 @@
 # Reproduce public and private
-python Ensemble1.py
-python Ensemble2.py
-python main.py --output_path ./
+PYTHONPATH=lib python Ensemble1.py
+# python Ensemble2.py
+# python main.py --output_path ./
